Clarify readme: .properties is a middleware function
When I first read it, I didn't notice the "s" of "Makes" and thought
this line was a reminder about how I have to configure my server.
Had almost suggested this package provides a helper for it, when
fortunately the example code made me realize this *is* that helper
function.